Add user management features and validations
Updated UserController with endpoints for user CRUD operations, including password changes. Enhanced UserProfile for mapping updates. Improved UserService with email checks and error handling. Overrode SaveChangesAsync in ErpDbContext for timestamp management. Added validators for user DTOs and introduced custom exceptions for better error handling. Updated UserAuthException status code and created new DTO classes for password changes and user updates.

+++ Erp.Application/Services/UserServices/UserService.cs
@@ -13,6 +13,7 @@ using Erp.Domain.DTOs.Auth;
 using Erp.Domain.CustomExceptions;
 using Erp.Infrastructure.Attributes.Caching;
 using Erp.Domain.Enums;
+using Microsoft.EntityFrameworkCore;
 
 namespace Erp.Application.Services.UserServices;
 
@@ -30,6 +31,11 @@ public class UserService : IUserService
 
 	public async Task<UserDto> CreateUserAsync(UserCreateDto userCreateDto)
 	{
+		var userExists = _db.Users.Any(x => x.Email == userCreateDto.Email);
+		if (userExists)
+		{
+			throw new UserEmailAlreadyExistsException();
+		}
 		var user = _mapper.Map<User>(userCreateDto);
 		user.PasswordHash = _passwordHasher.HashPassword(user, userCreateDto.Password);
 		await _db.Users.AddAsync(user);
@@ -48,6 +54,10 @@ public class UserService : IUserService
 	public async Task<UserDto> GetUserByIdAsync(Guid id)
 	{
 		var user = _db.Users.FirstOrDefault(x => x.Id == id);
+		if (user == null)
+		{
+			throw new NullValueException("Kullanıcı bulunamadı");
+		}
 		var dto = _mapper.Map<UserDto>(user);
 		return dto;
 	}
@@ -55,6 +65,10 @@ public class UserService : IUserService
 	public async Task HardDeleteUserAsync(Guid id)
 	{
 		var user = _db.Users.FirstOrDefault(x => x.Id == id);
+		if (user == null)
+		{
+			throw new NullValueException("Kullanıcı bulunamadı");
+		}
 		_db.Users.Remove(user);
 		await _db.SaveChangesAsync();
 	}
@@ -62,8 +76,40 @@ public class UserService : IUserService
 	public async Task SoftDeleteUserAsync(Guid id)
 	{
 		var user = _db.Users.FirstOrDefault(x => x.Id == id);
+		if (user == null)
+		{
+			throw new NullValueException("Kullanıcı bulunamadı");
+		}
 		user.IsDeleted = true;
 		_db.Users.Update(user);
 		await _db.SaveChangesAsync();
+	}
+
+	public async Task<UserDto> UpdateUserAsync(UserUpdateDto userUpdateDto, Guid userId)
+	{
+		var userEntity = await _db.Users.FirstOrDefaultAsync(x => x.Email == userUpdateDto.Email && x.Id != userId);
+		if (userEntity == null)
+		{
+			throw new UserEmailAlreadyExistsException();
+		}
+		var user = _mapper.Map(userUpdateDto, userEntity);
+		_db.Users.Update(user);
+		await _db.SaveChangesAsync();
+		var dto = _mapper.Map<UserDto>(user);
+		return dto;
+	}
+
+	public async Task<UserDto> ChangePasswordUserAsync(ChangePasswordUserDto changePasswordUserDto, Guid userId)
+	{
+		var userEntity = await _db.Users.FirstOrDefaultAsync(x => x.Id == userId);
+		if (userEntity == null)
+		{
+			throw new UserEmailAlreadyExistsException();
+		}
+		userEntity.PasswordHash = _passwordHasher.HashPassword(userEntity, changePasswordUserDto.Password);
+		_db.Users.Update(userEntity);
+		await _db.SaveChangesAsync();
+		var dto = _mapper.Map<UserDto>(userEntity);
+		return dto;
 	}
 }
